chore: finish type agumentation
