+++ packages/react/src/styled-system/register.ts
@@ -11,6 +11,7 @@ import type {
   Token as InternalToken,
   Tokens as InternalTokens,
 } from "./generated/token.gen"
+import type { RecipeDefinition, SlotRecipeDefinition } from "./recipe.types"
 
 export interface Register {}
 
@@ -45,3 +46,17 @@ export type Token = Register extends { token: infer T } ? T : InternalToken
 export type ColorPalette = Register extends { colorPalette: infer T }
   ? T
   : InternalColorPalette
+
+export type RecipeProps<T> = T extends keyof ConfigRecipes
+  ? ConfigRecipes[T]["__type"] & { recipe?: RecipeDefinition | undefined }
+  : { recipe?: RecipeDefinition | undefined }
+
+export type SlotRecipeProps<T> = T extends keyof ConfigSlotRecipes
+  ? ConfigSlotRecipes[T]["__type"] & {
+      recipe?: SlotRecipeDefinition | undefined
+    }
+  : { recipe?: SlotRecipeDefinition | undefined }
+
+export type SlotRecipeRecord<T, K> = T extends keyof ConfigRecipeSlots
+  ? Record<ConfigRecipeSlots[T], K>
+  : Record<string, K>
